Add file 4
This commit adds a new file, 4-user-promise.js, which contains the implementation of the signUpUser function. The function takes in a first name and last name as parameters and returns a resolved promise with an object syntax of {FirstName: value, LastName: value}. This implementation improves the codebase by providing a convenient way to create a resolved promise with the desired object structure.