Add ES6 class initialization function
This commit adds a new file, 1-make_classrooms.js, which exports a function called initializeRooms. This function imports the ClassRoom class from the 0-classroom file and initializes an array of ClassRoom objects with specified sizes. The sizes are provided as an array [19, 20, 34]. The function returns the array of initialized ClassRoom objects.