feat: Add SkyHighBuilding class that extends Building
The commit adds a new file, 6-sky_high.js, which exports a class called SkyHighBuilding. The SkyHighBuilding class extends the Building class and adds a new attribute, floors. It also overrides the evacuationWarningMessage method to return a specific message based on the number of floors.

Update of Building class was also made