Implement updateStudentGradeByCity function
This commit adds a new file, 4-update_grade_by_city.js, which exports a function called updateStudentGradeByCity. The function takes a list of students, a city, and an array of new grades as parameters. It filters the students based on the city and updates their grades according to the new grades array. If a student doesn't have a grade in the new grades array, their grade is set to 'N/A'. The function uses the filter and map array methods to achieve this.