Add createInt8TypedArray function
The commit adds a new file, 5-typed_arrays.js, which exports a function called createInt8TypedArray. This function creates a new ArrayBuffer with the specified length, creates a DataView to interact with the ArrayBuffer, and sets the Int8 value at the specified position. If the position is outside the range of the buffer, an error is thrown. The function returns the created DataView.