Add updateUniqueItems function to update quantities in a map
The commit adds a new file, 10-update_uniq_items.js, which exports a function called updateUniqueItems. This function accepts a map as an argument and updates the quantity of items with an initial quantity of 1 to 100. If the argument is not a map, it throws an error. This function is useful for updating quantities in a map of items.