Add weakMap and queryAPI functions
The commit adds a new file, 100-weak.js, which exports a const instance of WeakMap named `weakMap`. It also exports a new function named `queryAPI` that accepts an `endpoint` argument. The `queryAPI` function tracks the number of times it is called for each `endpoint` using the `weakMap`. If the number of queries is greater than or equal to 5, it throws an error with the message "Endpoint load is high". This implementation allows for efficient tracking and error handling of API queries.